+++ src/port_diff/extract.rs
@@ -33,7 +33,9 @@ impl<G: Graph> PortDiff<G> {
             .expect("not empty");
 
         // For each node that is an ancestor of two or more graphs...
-        for diff_ptr in GraphView::lowest_common_ancestors(&graphs) {
+        // TODO: this does not work
+        // for diff_ptr in GraphView::lowest_common_ancestors(&graphs) {
+        for diff_ptr in merged_graph.all_nodes() {
             // Check that its outgoing edges are compatible
             // (this must hold everywhere, but already holds elsewhere as the
             // set of outgoing edges in non-lca nodes remains unchanged).