+++ viewer/shared/src/model.rs
@@ -112,18 +112,16 @@ impl<G: pd::Graph> LoadedModel<G> {
     }
 }
 
-fn as_opt<V, E>(r: Result<Option<V>, E>) -> Option<Result<V, E>> {
-    match r {
-        Ok(None) => None,
-        Ok(Some(v)) => Some(Ok(v)),
-        Err(e) => Some(Err(e)),
+impl LoadedModel<StaticSizeCircuit> {
+    fn is_acyclic(&self) -> bool {
+        let node_ids = self
+            .selected_diffs
+            .iter()
+            .map(|diff| self.diff_id_to_ptr[diff.0 as usize]);
+        let diffs: Vec<_> = node_ids.map(|n| self.all_diffs.get_diff(n)).collect();
+        let circ = PortDiff::extract_graph(diffs).unwrap();
+        circ.is_acyclic()
     }
-}
-
-fn is_acyclic(diffs: Vec<PortDiff<StaticSizeCircuit>>) -> bool {
-    let circ = PortDiff::extract_graph(diffs).unwrap();
-    let mut cmds = circ.commands();
-    iter::from_fn(|| as_opt(cmds.try_next())).all(|x| x.is_ok())
 }
 
 impl Model {
@@ -159,16 +157,7 @@ impl Model {
         match self {
             Model::None => true,
             Model::Portgraph(model) => model.are_compatible(),
-            Model::Tket(model) => {
-                model.are_compatible() && {
-                    let node_ids = model
-                        .selected_diffs
-                        .iter()
-                        .map(|diff| model.diff_id_to_ptr[diff.0 as usize]);
-                    let diffs: Vec<_> = node_ids.map(|n| model.all_diffs.get_diff(n)).collect();
-                    is_acyclic(diffs)
-                }
-            }
+            Model::Tket(model) => model.are_compatible() && model.is_acyclic(),
         }
     }